Extract local CityGML construction source
